Use caret ^ dependencies instead of tilde ~ dependencies
I hope this may fix some problems with sourcemaps, by allowing to update
convert-source-map.

The decision to use ~ was explicit, and there was even a test in the
code which prevented the use of ^, added in ddfa437e344b42d1423ddd53714b344f77ee131b.
The rationale in 73689c7fc80c124143a577901d2b72476335b62b is:

    ^ operator in package.json is unnecessary and unfriendly to users

I disagree with that; let's hope I'm not wrong about that!