Fix sourcemap generation regarding member expressions
acorn-walk by default skips the property part of member expressions
when iterating. This meant that the properties were in the same node
as the base object when generating source maps.

E.g. in "this.foo.bar = 42;" the left part would not have separate
mapping entries, "this.foo.bar" would be mapped to the same node.